Add files via upload
This is the current bootable .bin file. Call it from grub to get it to boot. Otherwise, you can use QEMU with the argument -kernel and the filename to virtualize it directly. This is added, and will be updated, to allow those without the GCC i686 Cross Compiler to try the OS. I plan to add the .iso soon. Right now, it isn't a big deal, as it doesn't have a functional userspace, but later, I hope people will enjoy using the OS without building it themselves.